move: generator 패키지 분리
