+++ src/main/java/chess/domain/chessBoard/generator/PieceGenerator.java
@@ -1,4 +1,4 @@
-package chess.domain.chessBoard;
+package chess.domain.chessBoard.generator;
 
 import chess.domain.piece.Bishop;
 import chess.domain.piece.BlackPawn;
