+++ src/main/java/chess/domain/chessBoard/generator/SpaceGenerator.java
@@ -1,4 +1,6 @@
-package chess.domain.chessBoard;
+package chess.domain.chessBoard.generator;
+
+import chess.domain.chessBoard.Space;
 
 import java.util.List;
